cambiar primitivos por referencias en entidades

+++ previo2/src/main/java/com/example/demo/entities/FavoritoId.java
@@ -5,8 +5,8 @@ import java.util.Objects;
 
 public class FavoritoId implements Serializable {
 
-    private Long manga;
-    private Long usuario;
+    private Integer manga;
+    private Integer usuario;
 
     @Override
     public boolean equals(Object o) {
